requirements.txt
requests==2.23.0
requests-toolbelt==0.9.1
python-magic==0.4.15
beautifulsoup4==4.8.2